requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
edge-containers-cli==4.2.1
idna==3.10
Jinja2==3.1.4
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
natsort==8.4.0
platformdirs==4.3.6
polars-lts-cpu==1.12.0
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
shellingham==1.5.4
textual==0.85.1
typer==0.12.5
typing_extensions==4.12.2
uc-micro-py==1.0.3
urllib3==2.2.3
